feat: fetch options and info pages from API and show in footprint project type
